+++ cmds/link-testnet.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-symbol-bootstrap link -t target/testnet $1
+symbol-bootstrap link -t target/testnet-dual $1 $2 $3