+++ test/service/BootstrapUtils.test.ts
@@ -19,8 +19,9 @@ import { statSync } from 'fs';
 import 'mocha';
 import { it } from 'mocha';
 import { totalmem } from 'os';
-import { Account, Deadline, NetworkType, UInt64, VotingKeyLinkTransaction } from 'symbol-sdk';
+import { Account, Deadline, LinkAction, NetworkType, UInt64, VotingKeyLinkTransaction } from 'symbol-sdk';
 import { BootstrapUtils } from '../../src/service';
+import { CryptoUtils } from '../../src/service/CryptoUtils';
 import assert = require('assert');
 
 describe('BootstrapUtils', () => {
@@ -99,11 +100,48 @@ describe('BootstrapUtils', () => {
         expect(BootstrapUtils.toHex("5E62'990D'CAC5'BE8A")).to.be.eq("0x5E62'990D'CAC5'BE8A");
     });
 
+    it('BootstrapUtils.loadYaml', async () => {
+        expect(CryptoUtils.encryptedCount(BootstrapUtils.loadYaml('test/encrypted.yml', '1234'))).to.be.eq(0);
+
+        try {
+            BootstrapUtils.loadYaml('test/encrypted.yml', 'abc');
+            expect(1).eq(0);
+        } catch (e) {
+            expect(e.message).eq('Password is too short. It should have at least 4 characters!');
+        }
+
+        try {
+            BootstrapUtils.loadYaml('test/encrypted.yml', 'abcd');
+            expect(1).eq(0);
+        } catch (e) {
+            expect(e.message).eq('Cannot decrypt file test/encrypted.yml. Have you used the right --password param?');
+        }
+
+        try {
+            BootstrapUtils.loadYaml('test/encrypted.yml', '');
+            expect(1).eq(0);
+        } catch (e) {
+            expect(e.message).eq(
+                'File test/encrypted.yml seems to be encrypted but no password has been provided. Have you used the --password param?',
+            );
+        }
+
+        try {
+            BootstrapUtils.loadYaml('test/encrypted.yml', undefined);
+            expect(1).eq(0);
+        } catch (e) {
+            expect(e.message).eq(
+                'File test/encrypted.yml seems to be encrypted but no password has been provided. Have you used the --password param?',
+            );
+        }
+
+        expect(CryptoUtils.encryptedCount(BootstrapUtils.loadYaml('test/encrypted.yml', false))).to.be.eq(6);
+    });
+
     it('createVotingKeyTransaction v1 short key', async () => {
         const networkType = NetworkType.PRIVATE;
         const deadline = Deadline.createFromDTO('1');
         const voting = Account.generateNewAccount(networkType);
-        const currentHeight = UInt64.fromUint(10);
         const presetData = {
             networkType,
             votingKeyStartEpoch: 1,
@@ -113,7 +151,7 @@ describe('BootstrapUtils', () => {
 
         const transaction = BootstrapUtils.createVotingKeyTransaction(
             voting.publicKey,
-            currentHeight,
+            LinkAction.Link,
             presetData,
             deadline,
             maxFee,