+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.4.1] - Next
+
+**Milestone**: Hippopotamus(0.10.0.5)
+
+ Package  | Version  | Link
+---|---|---
+Symbol Bootstrap | v0.4.1 | [symbol-bootstrap](https://www.npmjs.com/package/symbol-bootstrap)
+
+- Improved --password. It's only required when private keys need to be read.
+- Added `database` service to server and broker `depends_on` compose services.
+- Fixed `link --unlink` command for Voting Key Link transactions.
+- Added multisig account validation to `link` and `supernode` commands.
+- Added `CONTROLLER_PUBLIC_KEY` to Supernode's agent configuration
+- Upgraded Symbol Rest to version 2.3.1.
+
 ## [0.4.0] - Jan-14-2020
 
 **Milestone**: Hippopotamus(0.10.0.5)